Use openssl base64 for gcr_helper (#40)
The base64 command is using the `-w` flag to disable wrapping. On the Mac the equivalent flag is `-b`.  There is no common flag that will disable wrapping.

One option is to use `openssl base64 -a -A`. I'm not sure that's the right answer, but I'm opening a PR to start the discussion.