buildah: successful build template
Recent releases of buildah require newer than repo and newer than latest
tag of runc, so build it from source.
But allow for build-args to the Dockerfile, so you could specify
revisions of runc and buildah as needed.

Building as a limited user inside the container, using userns, is
largely there, but the shadow-utils of centos7 are not. So sticking with
building as uid=0 inside the container for now.

Adding the `--layers` flag with leverage build cache in the volume, if
present.

Signed-off-by: Vincent Batts <vbatts@hashbangbash.com>